backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
httpx==0.25.2
apscheduler==3.10.4
pydantic==2.5.0
python-dotenv==1.0.0
